requirements.txt
# Core dependencies
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
pyyaml>=6.0
requests>=2.31.0
aiohttp>=3.8.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# AI/ML dependencies
openai>=1.0.0
anthropic>=0.7.0
transformers>=4.30.0
torch>=2.0.0

# Social media APIs
tweepy>=4.14.0
praw>=7.7.0
httpx>=0.24.0

# Utilities
typer>=0.9.0
colorama>=0.4.6
cachetools>=5.3.0
beautifulsoup4>=4.12.0
markdown>=3.4.0
emoji>=2.2.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0
